Moved testing code to gosub_html5 crate
