+++ crates/gosub_html5/Cargo.toml
@@ -14,14 +14,20 @@ lazy_static = "1.5"
 thiserror = "2.0.7"
 url = { version = "2.5.4", features = [] }
 log = { version = "0.4.22", features = [] }
+nom = "7.1.3"
+nom_locate = "4.2.0"
+regex = "1"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = { version = "1.0", features = ["preserve_order"] }
+serde_derive = "1.0"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 ureq = "2.12.1"
 
 [dev-dependencies]
-gosub_testing = { path = "../gosub_testing" }
 test-case = "3.3.1"
 criterion = { version = "0.5.1", features = ["html_reports"] }
+
 
 [features]
 debug_parser = []
