+++ crates/gosub_html5/src/testing.rs
@@ -2,5 +2,5 @@
 pub mod tokenizer;
 pub mod tree_construction;
 
-pub const FIXTURE_ROOT: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/../gosub_html5/tests/data/html5lib-tests",);
+pub const FIXTURE_ROOT: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/./tests/data/html5lib-tests",);
 pub const TREE_CONSTRUCTION_PATH: &str = "tree-construction";
