+++ crates/gosub_html5/src/testing/tree_construction.rs
@@ -1,10 +1,10 @@
 pub mod fixture;
 mod generator;
-pub(crate) mod parser;
+pub mod parser;
 pub mod result;
 
 use generator::TreeOutputGenerator;
-use gosub_html5::node::{HTML_NAMESPACE, MATHML_NAMESPACE, SVG_NAMESPACE};
+use crate::node::{HTML_NAMESPACE, MATHML_NAMESPACE, SVG_NAMESPACE};
 use gosub_shared::byte_stream::{ByteStream, Config, Encoding, Location};
 use gosub_shared::document::DocumentHandle;
 use gosub_shared::node::NodeId;
