--- crates/gosub_testing/src/lib.rs
@@ -1,6 +1,0 @@
-//! Testing functionality
-//!
-//! This crate supplies testing functionality to test all the gosub crates.
-//!
-
-pub mod testing;
