+++ src/bin/html5-parser-test.rs
@@ -5,8 +5,8 @@ use gosub_html5::document::fragment::DocumentFragmentImpl;
 use gosub_html5::parser::Html5Parser;
 use gosub_shared::traits::config::{HasCssSystem, HasDocument, HasHtmlParser};
 use gosub_shared::types::Result;
-use gosub_testing::testing::tree_construction::fixture::{fixture_root_path, read_fixture_from_path};
-use gosub_testing::testing::tree_construction::Harness;
+use gosub_html5::testing::tree_construction::fixture::{fixture_root_path, read_fixture_from_path};
+use gosub_html5::testing::tree_construction::Harness;
 use std::io::Write;
 use std::path::{Path, PathBuf};
 use walkdir::WalkDir;
